_BatchConfig: optimize '_normalize_csnames()'
Change 'allcsnames' from list to set. Duplicate values can not be added
to a set, which makes it faster than searching a list for duplicate
value, before adding.

Uppercase all values in 'csnames ' once, instead of 2 times.

Line 271: "return [csname for csname in allcsnames if csname in csnames]"
is useless, because we raise exception on line 214 if a 'csname' is not in
'allcsnames'. Thus, just return 'csnames'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>