_BatchConfig: revert sets to lists
Partially revert commits:
  d916e0a "_BatchConfig: optimize '_normalize_csnames()'"
  87f63bc "_BatchConfig: optimize'_normalize_pcsnames()'"

Revert sets to lists, because lists keep the order that the elements are added.
Sets are more optimal in this situation, except for when "all" is passed. In
this case all available package/requestable C-states names are returned, and it
is cleaner to keep them in order. For example:
 "PC0", "PC2", "PC6N", "PC6R", "unlimited"

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>