_ReportBase: change 'self._smry_metrics' to a dictionary
Different results may contain different metrics. This patch changes
'self._smry_metrics' to a dictionary so that each result can have its
own list of metrics for which summaries should be calculated.

In turn this facilitates a future change which will allow metric tabs to
be generated for metrics which are not present in all results data.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>