_ReportBase: generate metric tabs when one or more results have data
This patch changes the way that tab parameters ('xaxes', 'yaxes', 'hist'
and 'chist') are instantiated to allow diagrams and tabs to be generated
if one or more results have data for them. Without this patch, all
results must have data to be included in a tab for it to be generated.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>