_FTrace: start the trace_pipe reader process only when required
In some cases, _FTrace module is used to only configure the tracer, and
it is not desirable to spawn the reader process; e.g. with eBPF helpers
which handle the trace log parsing themselves. For this purpose, modify
the trace_pipe reader process spawning to only happen on the first call
to the 'getlines()' method.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>