exercisesut: remove hard-coded tool names
'wult', 'ndl', and 'stats-collect' all have 'ToolInfo' modules which
provide a public constant 'TOOLNAME', representing the current name of
each tool accordingly.

This patch removes hard-coded instances of the tool names and replaces
them with references to the public constants offered by the tools.

This would make it easier in future if any of the tools need to be
renamed.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>