_ReportBase: remove 'base metric' logic
'stats-collect v1.0.18' adds automatic scaling of data which can be
scaled to take advantage of 'plotly' SI-unit prefix handling. It also
formats the data in hover text accordingly.

The 'base metric' logic in 'ReportBase' was implemented to provide these
two functions. Therefore this patch removes the 'base metric' logic
completely as it is no longer needed.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>