PbeDefs: indicate that 'pbe' defs are part of the 'wult' project
Fix 'PbeDefs' so that it indicates that it is part of the 'wult'
project. Specifically, this means that 'pbe' data files (such as defs
files) can be pointed to using the 'WULT_DATA_PATH' environment
variable, rather than introducing a separate one just for 'pbe'.

This is more consistent with how the 'wult' and 'ndl' tools expose the
data path.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>