PbeRunner: enable 'MeasuredCPU' tab in HTML reports
Set the number of the CPU given to the pbe driver (which will be
responsible for sending interrupts to wake all other CPUs) as the
'Measured CPU'. This means that pbe results will keep a record of which
CPU was used as the 'main' CPU within the statistics result directory.

In turn, this is also used automatically during report generation to add
the 'Measured CPU' turbostat tab.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>