_Common: fix 'wult/ndl scan' crashing if one device is not supported
Fix an issue where 'wult scan' or 'ndl scan' errors as soon as one
device is not supported.

This issue can result in the following scenario:
1. The user runs a kernel which is too old to support the 'wult' driver
   but not too old to support the bpf helpers.
2. When they run 'wult scan', 'wult' errors out saying that the 'wult'
   driver is not available on their kernel but it does not let them know
   that they can still use the bpf methods.

Add 'ErrorNotSupported' to the exceptions which can be raised while
checking the deployment of devices so that a device not being supported
does not prevent the user from finding out that other methods are
supported.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>