wult: drivers: move initialization to armer kthread
Move most of the initialization to the armer kthread, to make sure it is done
on the measured CPU. At the moment it is important for C-states initialization
(intel_cstate_init()), because different CPUs may have different C-state
ounters supported (e.g., E-cores do not support CC7). But there may be other
parts of the initialization that should be done on the measured CPU, so just do
it all there.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>