exercise-sut: do not disable 'POLL' state
Always keep POLL enabled, even when '--only-one-cstate' is used. In kernel in
the 'menu' governor there is a shortcut for the case when there is only one
C-state enabled. We do not want to hit this special code path every time we
mwasure C-states. Instead, we want to mostly hit the standard code path, which
is run when there is more than one C-state.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>