exercise-sut: fix async commands being stuck
Sometimes creating reports with '--jobs' option
was getting stuck. This happens if subprocess,
e.g. stats-collect, prints lot of information to
stdout. The output is piped to stdout. The pipe
has maximum limit of characters it will receive
before it will block the process. The process is
put to sleep until the data is fetched from the
pipe.

To fix this, we need to use wait() methode instead
of poll(). Using wait() will run separate thread
to fetch data from the pipe. This will prevent
the thread being stuck.

Signed-off-by: Antti Laakso <antti.laakso@intel.com>