_BatchConfig: do not assume unique property names
From exercise-sut point of view, property is what is defined in 'PROP_INFOS'
dictionary.

Frpm papclibs point of view, property is what is defined in 'PState' or
'CState' module.

They property names mostly match, so '_BatchConfig' does not really distinguish
between the two different property definitions. It assumes that whatever is in
'PROP_INFO' will be found in one of 'PState' or 'CState' modules.

This is a problem for the "governor" property, because it exists in both
'PState' and 'CState' modules. Currently 'exercise-sut' only supports CPU
frequency governor, but adding idle governor support is problematic because of
this.

Refactor the code so that the 'PROP_INFOS' also points to the class of the
property and its name in the class.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>