wult: driver: remove unused function "set_bool"
If the Linux kernel on the SUT is compiled with CONFIG_WERROR kernel
configuration set to true, then the "wult deploy" command fails to
execute. The command fails to execute because the "set_bool" function
has been defined but not used (-Werror=unused-function), and all warings
are being treated as errors when CONFIG_WERROR is enabled.

Fix the error by removing the unused function.

Signed-off-by: Harry Roshan <harryroshan1997@gmail.com>