exercise-sut: rename "__reportid__" to "{REPORTID}"
Use "{}" for the substitute markers, instead of "____". Just a personal
preference, somewhat similar to f-strings, more intuitive to me. Also it
refers to option value, as printed in the help text and documentation.

E.g., the documentation talks about '--repoitid REPORTID', so "{REPORTID}"
in capital letters correspoinds to the 'REPORTID' part.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>